created dispense controller and routes
