+++ src/routes/medicationRoutes.js
@@ -8,27 +8,10 @@ const {
   createMedication,
   updateMedication,
   deleteMedication,
-  dispenseMedication,
-  getDispenseLogs,
 } = require("../controllers/medicationController");
 
 const router = express.Router();
 
-// POST /inventory/dispense Dispense medication
-router.post(
-  "/dispense",
-  authenticate,
-  roleMiddleware(["admin", "clerk", "inventoryManager"]),
-  dispenseMedication,
-);
-
-// Dispense Log Stats
-router.get(
-  "/dispense-logs",
-  authenticate,
-  roleMiddleware(["admin", "clerk", "inventoryManager"]),
-  getDispenseLogs,
-);
 // GET /inventory: Get all medications
 router.get(
   "/",
